% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_debajyoti.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot accident locations on map}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{State Id. This parameter isn't initialisied by default. Handles integers, floats as well as strings.}

\item{year}{Year (YYYY). This parameter isn't initialisied by default. Handles integers, floats as well as strings.}
}
\value{
map plot with rectangular markers for accident locations. Returns message if no accidents are available to be plotted.
}
\description{
Read fars data from file in current working directory and plot accident location for given state during the given year,
within a fixed maximum threshold for longitude (900) and latitude (90). Throws error if invalid values are provided for _year_ and _state_.
}
\examples{
fars_map_state(1,2013)
}

