% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_debajyoti.R
\name{fars_read}
\alias{fars_read}
\title{Read fars data}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{Name of input file (String). This parameter isn't initialisied by default.}
}
\value{
This function returns a dplyr dataframe
}
\description{
Check if input file exists in current working directory and if found, read data from it.
If file isn't found, throws error message and stops executing the function.
}

