% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_debajyoti.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read fars data based on input year}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{Year (YYYY). This parameter isn't initialisied by default. Handles integers, floats as well as strings.}
}
\value{
This function returns a subset of fars data containing only the year and month values in long format
}
\description{
Generates string name from given input _year_,
check if a matching input file exists in current working directory and if found, reads the data.
If matching file is not found in the current working directory, it throws an __invalid year__ error.
}
\examples{
fars_read_years(2013)
fars_read_years("2013")
}

