% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_debajyoti.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize fars data by Year}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{Year (YYYY). This parameter isn't initialisied by default. Handles integers, floats as well as strings.}
}
\value{
This functions returns a summarized dataframe giving count of instances per month of the year in a wide format. It returns one row for each month and one column per year in the data.
}
\description{
Try to read fars data in current working directory and if successful, summarize data by count of instances per month.
}
\examples{
\dontrun{
fars_summarize_years(2013)
fars_summarize_years("2013")
}
}

