% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_debajyoti.R
\name{make_filename}
\alias{make_filename}
\title{Generates a filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{Year (YYYY). This parameter isn't initialisied by default. Handles integers, floats as well as strings.}
}
\value{
This function returns a character vector to serve as a filename for an archived csv file.
}
\description{
Take _year_ value as input, converts it into an integer, and generates a character vector containing a formatted combination
of text and provided variable values.
}
\examples{
\dontrun{
make_filename(2013)
make_filename("2013")
make_filename("2013.076")
}
}

